failed attempt to use etherscan API to get pending TXs
